Initial user facing API, further performance improvements
