+++ core/licensing.cpp
@@ -3,9 +3,6 @@
 // License: specified in project_root/LICENSE.txt
 #include "husarnet/licensing.h"
 
-#include <algorithm>
-#include <fstream>
-
 #include <sodium.h>
 
 #include "husarnet/ports/port_interface.h"
