+++ core/ports/esp32/user_interface.cpp
@@ -1,16 +1,55 @@
 // Copyright (c) 2024 Husarnet sp. z o.o.
 // Authors: listed in project_root/README.md
 // License: specified in project_root/LICENSE.txt
-#include <husarnet/husarnet_manager.h>
+#include "user_interface.h"
 
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 
-extern "C" void run_husarnet(TaskHandle_t* husarnet_task_handle)
+// C++ API
+
+HusarnetClient::HusarnetClient()
 {
-  auto* manager = new HusarnetManager();
+  husarnetManager = new HusarnetManager();
+}
+
+HusarnetClient::~HusarnetClient()
+{
+  delete husarnetManager;
+}
+
+void HusarnetClient::start()
+{
+  void* manager = static_cast<void*>(husarnetManager);
 
   xTaskCreate([](void* manager) {
     ((HusarnetManager*)manager)->runHusarnet();
-  }, "husarnet_task", 16384, manager, 7, husarnet_task_handle);
+  }, "husarnet_task", 16384, manager, 7, NULL);
+}
+
+void HusarnetClient::join(const char* hostname, const char* joinCode)
+{
+  husarnetManager->joinNetwork(joinCode, hostname);
+}
+
+HusarnetManager* HusarnetClient::getManager()
+{
+  return husarnetManager;
+}
+
+// C API
+
+HusarnetClient* husarnet_init()
+{
+  return new HusarnetClient();
+}
+
+void husarnet_start(HusarnetClient* client)
+{
+  client->start();
+}
+
+void husarnet_join(HusarnetClient* client, const char* hostname, const char* joinCode)
+{
+  client->join(hostname, joinCode);
 }
