+++ platforms/esp32/flash_monitor.sh
@@ -5,4 +5,4 @@
 set -e
 
 platforms/esp32/idf.py -p "$1" -b 921600 flash
-tio -b 921600 "$1"
+tio -b 115200 "$1"
