+++ platforms/esp32/main/CMakeLists.txt
@@ -1,3 +1,3 @@
-idf_component_register(SRCS "main.c"
+idf_component_register(SRCS "wifi.c" "main.c"
                     INCLUDE_DIRS ""
                     REQUIRES husarnet_core lwip esp_hw_support nvs_flash esp_netif esp_wifi)
