+++ platforms/esp32/main/wifi.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void wifi_init();
